+++ src/compas_ui/objects/networkobject.py
@@ -60,6 +60,10 @@ class NetworkObject(Object):
         self.item = network
 
     @property
+    def view_network(self):
+        return self.view_item
+
+    @property
     def anchor(self):
         return self._anchor
 
@@ -101,8 +105,9 @@ class NetworkObject(Object):
     @property
     def node_xyz(self):
         origin = Point(0, 0, 0)
-        nodes = list(self.network.nodes())
-        xyz = self.network.nodes_attributes(['x', 'y', 'z'], keys=nodes)
+        view_network = self.view_network
+        nodes = list(view_network.nodes())
+        xyz = view_network.nodes_attributes(['x', 'y', 'z'], keys=nodes)
 
         stack = []
         if self.scale != 1.0:
@@ -113,7 +118,7 @@ class NetworkObject(Object):
             stack.append(R)
         if self.location != origin:
             if self.anchor is not None:
-                xyz = self.network.node_attributes(self.anchor, 'xyz')
+                xyz = view_network.node_attributes(self.anchor, 'xyz')
                 point = Point(* xyz)
                 T1 = Translation.from_vector(origin - point)
                 stack.insert(0, T1)