+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `use_tab` option to `compas_ui.rhino.forms.SettingsForm`.
 * Added `export` and `remove` buttons to `compas_ui.rhino.forms.SceneForm`.
 * Added `compas_ui.rhino.forms.ToolbarForm`.
+* Added `frame` to `compas_ui.objects.Object`.
 
 ### Changed